+++ app/src/main/java/com/sakal/playlistmaker/creator/Creator.kt
@@ -6,16 +6,16 @@ import androidx.appcompat.app.AppCompatActivity
 import com.sakal.playlistmaker.ApiConstants
 import com.sakal.playlistmaker.BuildConfig
 import com.sakal.playlistmaker.Constants
-import com.sakal.playlistmaker.player.data.PlayerRepositoryImpl
+import com.sakal.playlistmaker.player.data.impl.PlayerRepositoryImpl
 import com.sakal.playlistmaker.player.domain.api.PlayerInteractor
 import com.sakal.playlistmaker.player.domain.api.PlayerRepository
 import com.sakal.playlistmaker.player.domain.impl.PlayerInteractorImpl
-import com.sakal.playlistmaker.search.data.SearchRepoImpl
+import com.sakal.playlistmaker.search.data.impl.SearchRepoImpl
 import com.sakal.playlistmaker.search.data.network.ApiService
 import com.sakal.playlistmaker.search.data.preferences.LocalStorage
 import com.sakal.playlistmaker.search.domain.HistoryDataStore
 import com.sakal.playlistmaker.search.domain.SearchInteractor
-import com.sakal.playlistmaker.search.domain.SearchInteractorImpl
+import com.sakal.playlistmaker.search.domain.impl.SearchInteractorImpl
 import com.sakal.playlistmaker.search.domain.SearchRepo
 import com.sakal.playlistmaker.search.ui.view_model.SearchViewModelFactory
 import com.sakal.playlistmaker.settings.data.impl.SettingsRepositoryImpl
